requirements.txt
langchain
langchain-core
langchain-groq
pymongo
python-dotenv
